pkgs/finance-tools/pyproject.toml
[project]
name = "finance-tools"
version = "0.1.12"
description = "Personal finance ingestion tools"
authors = [{name = "Inkpot Monkey", email = "inkpotmonkey@example.com"}]
dependencies = [
    "beangulp",
    "beancount",
    "fava",
    "python-dateutil",
    "openai",
    "scikit-learn",
    "litellm",
]
requires-python = ">=3.10"

[project.scripts]
finance-ingest-internal = "ingest:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["ingest", "config"]
packages = ["importers", "hooks", "finance_extensions"]

[tool.setuptools.package-data]
"finance_extensions.templates" = ["*.html"]
